+++ index.js
@@ -5,6 +5,8 @@ var reduceCSSCalc = require("reduce-css-calc")
 var helpers = require("postcss-message-helpers")
 var postcss = require("postcss")
 
+var CONTAINS_CALC = /calc\(.*\)/
+
 /**
  * PostCSS plugin to reduce calc() function calls.
  */
@@ -12,8 +14,10 @@ module.exports = postcss.plugin("postcss-calc", function(options) {
   options = options || {}
   var precision = options.precision
   var preserve = options.preserve
+  var warnWhenCannotResolve = options.warnWhenCannotResolve
 
-  return function(style) {
+  return function(style, result) {
+
     style.walkDecls(function transformDecl(decl) {
       if (!decl.value || decl.value.indexOf("calc(") === -1) {
         return
@@ -21,6 +25,11 @@ module.exports = postcss.plugin("postcss-calc", function(options) {
 
       helpers.try(function transformCSSCalc() {
         var value = reduceCSSCalc(decl.value, precision)
+
+        if (warnWhenCannotResolve && CONTAINS_CALC.test(value)) {
+          result.warn("Could not reduce expression: " + decl.value,
+            {plugin: "postcss-calc", node: decl})
+        }
 
         if (!preserve) {
           decl.value = value